did_finder_rucio/requirements.txt
rucio-clients
pika

